Move timestamp function to platform files per review feedback
Move SV_TimestampMilliseconds from sv_demo_misc.c into sv_sys_unix.c
and sv_sys_win.c as Sys_TimestampMilliseconds, following the existing
convention for platform-specific functions.
